radeonsi: Use CE for all descriptors

+++ src/gallium/drivers/radeonsi/si_state.h
@@ -197,8 +197,8 @@ struct si_descriptors {
 	unsigned ce_offset;
 
 	/* The i-th bit is set if that element is enabled (non-NULL resource). */
-	uint64_t enabled_mask;
-	uint64_t dirty_mask; /* needed by the CE for partial updates */
+	uint32_t enabled_mask;
+	uint32_t dirty_mask; /* needed by the CE for partial updates */
 
 	/* The shader userdata offset within a shader where the 64-bit pointer to the descriptor
 	 * array will be stored. */
